Renamed output dto file to match the convention
