+++ ghost/ghost/src/http/controllers/snippet.dto.output.ts
